Restructured test folders for better organization.
